BE Just travels backend  backup code

+++ lg-lib/source/db_operations/travel/deletebyid.hpp
@@ -82,7 +82,7 @@ private:
     QString sqlStm = "UPDATE drivers SET blocked = 0 WHERE  id = :id ";
 
     std::map<QString, QVariant> binds;
-    binds.insert(Burden(":id", QVariant(jo_["destiny"]["id_dri_o"])) );
+    binds.insert(Burden(":id", QVariant(jo_["destiny"]["id_dri_o"].toInt())) );
 
     return  db_.update(sqlStm, binds);
   }
@@ -92,7 +92,7 @@ private:
     QString sqlStm = "UPDATE vehicles SET blocked = 0 WHERE  id = :id ";
 
     std::map<QString, QVariant> binds;
-    binds.insert(Burden(":id", QVariant(jo_["destiny"]["id_veh_o"])) );
+    binds.insert(Burden(":id", QVariant(jo_["destiny"]["id_veh_o"].toInt())) );
 
     return  db_.update(sqlStm, binds);
   }
