+++ lg-lib/source/db_operations/travel/insert.hpp
@@ -49,7 +49,7 @@ private:
     QString sqlStm = "SELECT 1 FROM drivers WHERE blocked = 1 AND id = :id ";
 
     std::map<QString, QVariant> binds;
-    binds.insert(Burden(":id", QVariant(jo_["destiny"]["id_dri"])) );
+    binds.insert(Burden(":id", QVariant(jo_["destiny"]["id_dri"].toInt())) );
 
     return  db_.search(sqlStm, binds).next();
   }
@@ -59,7 +59,7 @@ private:
     QString sqlStm = "SELECT 1 FROM vehicles WHERE blocked = 1 AND id = :id ";
 
     std::map<QString, QVariant> binds;
-    binds.insert(Burden(":id", QVariant(jo_["destiny"]["id_veh"])) );
+    binds.insert(Burden(":id", QVariant(jo_["destiny"]["id_veh"].toInt())) );
 
     return  db_.search(sqlStm, binds).next();
   }
@@ -73,7 +73,7 @@ private:
       " ( :id_client, :sta_date )";
 
     std::map<QString, QVariant> binds;
-    binds.insert(Burden(":id_client", QVariant(jo_["id_cli"])) );
+    binds.insert(Burden(":id_client", QVariant(jo_["id_cli"].toInt())) );
     binds.insert(Burden(":sta_date",  QVariant(jo_["sta_date"])) );
 
     return  db_.create(sqlStm, binds);
@@ -90,8 +90,8 @@ private:
        " , :street, :house_nro, :post_code) ";
 
     std::map<QString, QVariant> binds;
-    binds.insert(Burden( ":id_driver",  QVariant(jo_["destiny"]["id_dri"]) ));
-    binds.insert(Burden( ":id_vehicle", QVariant(jo_["destiny"]["id_veh"]) ));
+    binds.insert(Burden( ":id_driver",  QVariant(jo_["destiny"]["id_dri"].toInt()) ));
+    binds.insert(Burden( ":id_vehicle", QVariant(jo_["destiny"]["id_veh"].toInt()) ));
     binds.insert(Burden( ":arrival_date",
           QVariant(jo_["destiny"]["arr_date"]) ));
 
@@ -127,7 +127,7 @@ private:
     QString sqlStm = "UPDATE drivers SET blocked = 1 WHERE  id = :id ";
 
     std::map<QString, QVariant> binds;
-    binds.insert(Burden(":id", QVariant(jo_["destiny"]["id_dri"])) );
+    binds.insert(Burden(":id", QVariant(jo_["destiny"]["id_dri"].toInt())) );
 
     return  db_.update(sqlStm, binds);
   }
@@ -137,7 +137,7 @@ private:
     QString sqlStm = "UPDATE vehicles SET blocked = 1 WHERE  id = :id ";
 
     std::map<QString, QVariant> binds;
-    binds.insert(Burden(":id", QVariant(jo_["destiny"]["id_veh"])) );
+    binds.insert(Burden(":id", QVariant(jo_["destiny"]["id_veh"].toInt())) );
 
     return  db_.update(sqlStm, binds);
   }
